Added admin and guest roles

+++ controllers/featured_controller.js
@@ -7,15 +7,19 @@ const { user, program, feature } = db
 featuredPrograms.get('/', async (req, res) => {
   try {
     const foundFeatured = await feature.findAll({
-      order: [['featured_id', 'ASC']],
+      order: [['createdAt', 'ASC']],
       include: [
         {
-          model: user,
-          as: 'user'
-        },
-        {
           model: program,
-          as: 'programs'
+          as: 'program',
+          attributes: ['title', 'description'],
+          include: [
+            {
+              model: user,
+              as: 'author',
+              attributes: ['user_id', 'username']
+            }
+          ]
         }
       ]
     })
@@ -27,11 +31,21 @@ featuredPrograms.get('/', async (req, res) => {
 
 // CREATE A FEATURED PROGRAM
 featuredPrograms.post('/', (req, res) => {
+
+  if (req.currentUser === null || req.currentUser.role !== 'admin') {
+    return res.status(403).json({ message: 'You must be an admin to add programs to the featured list' })
+  }
+
   res.send('Got a POST request')
 })
 
 // UPDATE A FEATURED PROGRAM
 featuredPrograms.put('/:id', async (req, res) => {
+
+  if (req.currentUser === null || req.currentUser.role !== 'admin') {
+    return res.status(403).json({ message: 'You must be an admin to add programs to modify the featured list' })
+  }
+
   try {
     const updatedFeatured = await feature.update(req.body, {
       where: {
@@ -48,6 +62,11 @@ featuredPrograms.put('/:id', async (req, res) => {
 
 // DELETE A FEATURED PROGRAM
 featuredPrograms.delete('/:id', async (req, res) => {
+
+  if (req.currentUser === null || req.currentUser.role !== 'admin') {
+    return res.status(403).json({ message: 'You must be an admin to remove programs from the featured list' })
+  }
+
   try {
     const deletedFeatured = await feature.destroy({
       where: {
